+++ fj-db-extreme/src/test/java/org/forumj/dbextreme/db/service/test/user/tests/ReadUserTest.java
@@ -76,7 +76,7 @@ public class ReadUserTest {
       st.setInt(18, user.getShowBithday() ? 1 : 0);
       st.setInt(19, user.getLanguge().getCode());
       st.setInt(20, user.getHideIp() ? 1 : 0);
-      st.setInt(21, user.getView());
+      st.setLong(21, user.getView());
       st.setInt(22, user.getPp());
       st.setInt(23, user.getPt());
       st.setString(24, user.getAvatar());
